// src/app/models/eventReport.ts
export interface EventReport {
    id: number;
    //reportUUID: string;
    im1: string;
    im2: ImageBitmap;
    im3: ImageBitmap;
    im4: ImageBitmap;
    im5: ImageBitmap;
    im6: ImageBitmap;
    im7: ImageBitmap;
    im8: ImageBitmap;
    im9: ImageBitmap;
    im10: ImageBitmap;
    drugs: string;
    anest: string;
    assistant: string;
    scope: string;
    phar: string;
    esop: string;
    stom: string;
    duod: string;
    urease: string;
    biopsy: string;
    hystoResults: string;
    prep: string;
    quality: string;
    colo: string;
    conc1: string;
    conc2: string;
    conc3: string;
    conc4: string;
    conc5: string;
    conc6: string;
    complications: string;
    event: number;
}